+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 
 **Bugfixes**
 - Fixed crash due to broken EMI compat.
+- [FORGE] Fixed lockdown effect icon not displaying when rejoining world.